feat: 검색 결과 컴포넌트 분리
- FAQList에서 검색 결과를 표시하는 코드를 SearchResult 컴포넌트로 분리
- useEffect 를 사용하여  검색 결과를 SearchContext 에 저장하고, SearchResult 컴포넌트가 이를 구독하여 값이 업데이트될 수 있도록 함
