+++ components/FAQList/FAQList.module.scss
@@ -2,30 +2,6 @@
 @use '@/styles/breakpoints' as breakpoint;
 @use '@/styles/colors' as color;
 
-.searchResult {
-  display: flex;
-  justify-content: space-between;
-  margin: $px-md 0;
-  font-size: $heading-info;
-}
-
-.searchResultText {
-  font-weight: bold;
-}
-
-.searchResultCount {
-  margin-left: 4px;
-  font-weight: 700;
-  color: color.$mint-900;
-}
-
-.resetButton {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  font-size: 16px;
-}
-
 .faqList {
   display: flex;
   flex-direction: column;
