+++ components/FAQList/index.tsx
@@ -1,13 +1,11 @@
 'use client';
 
+import { useEffect } from 'react';
 import FAQAccordion from '@/components/FAQAccordion';
-import ResetIcon from '@/assets/icons/reset_icon.svg';
-import SubCategoryTabList from '@/components/SubCategoryTabList';
 import NoDataIcon from '@/assets/icons/no_data_icon.svg';
 import { useFaqList } from '@/services/useFAQService';
 import { useTabContext } from '@/hooks/useTabContext';
 import { useSearchContext } from '@/hooks/useSearchContext';
-import { useFormContext } from 'react-hook-form';
 import { CategoryType, SubCategoryType } from '@/types';
 import styles from './FAQList.module.scss';
 import Loading from '@/components/common/Loading';
@@ -17,10 +15,10 @@ type FAQListProps = {
 };
 
 export default function FAQList({ category }: FAQListProps) {
-  const { query, setQuery } = useSearchContext();
-  const { reset } = useFormContext();
+  const { query, setSearchResult } = useSearchContext();
   const { activeTab } = useTabContext();
   const subCategory = activeTab as SubCategoryType;
+
   const { data, fetchNextPage, hasNextPage, isFetchingNextPage, isFetching } =
     useFaqList({
       category,
@@ -29,31 +27,14 @@ export default function FAQList({ category }: FAQListProps) {
       query,
     });
 
-  const handleResetSearch = () => {
-    setQuery('');
-    reset();
-  };
+  useEffect(() => {
+    setSearchResult(data?.pages[0].items.length);
+  }, [data?.pages[0].items.length]);
 
   const noResult = query && data?.pages[0].items.length === 0;
 
   return (
     <>
-      {query && (
-        <div className={styles.searchResult}>
-          <p className={styles.searchResultText}>
-            검색 결과 총
-            <span className={styles.searchResultCount}>
-              {data?.pages[0].items.length}
-            </span>
-            건
-          </p>
-          <button onClick={handleResetSearch} className={styles.resetButton}>
-            <ResetIcon width={24} height={24} />
-            검색 초기화
-          </button>
-        </div>
-      )}
-      <SubCategoryTabList category={category} />
       <ul className={`${styles.faqList} ${noResult && styles.noResult}`}>
         {noResult && (
           <div className={styles.noResultContent}>
